restructure full filesystem
