+++ .gitignore
@@ -108,3 +108,4 @@ ecosystem.config.js
 gcp-key.json
 database/pgdata
 database/pgadmin
+/dev
